--- a/perf_test/sparse/KokkosSparse_sort_crs.cpp
+++ b/perf_test/sparse/KokkosSparse_sort_crs.cpp
@@ -35,8 +35,7 @@ void print_options() {
   std::cerr << perf_test::list_common_options();
 
   std::cerr << "\t[Required] --mtx <path> :: matrix to sort\n";
-  std::cerr
-      << "\t[Optional] --repeat      :: how many times to repeat sorting\n";
+  std::cerr << "\t[Optional] --repeat      :: how many times to repeat sorting\n";
 }
 
 int parse_inputs(LocalParams& params, int argc, char** argv) {
@@ -44,8 +43,7 @@ int parse_inputs(LocalParams& params, int argc, char** argv) {
     if (perf_test::check_arg_str(i, argc, argv, "--mtx", params.mtxFile)) {
       ++i;
     } else {
-      std::cerr << "Unrecognized command line argument #" << i << ": "
-                << argv[i] << std::endl;
+      std::cerr << "Unrecognized command line argument #" << i << ": " << argv[i] << std::endl;
       print_options();
       return 1;
     }
@@ -54,8 +52,7 @@ int parse_inputs(LocalParams& params, int argc, char** argv) {
 }
 
 template <typename exec_space>
-void run_experiment(int argc, char** argv,
-                    const CommonInputParams& common_params) {
+void run_experiment(int argc, char** argv, const CommonInputParams& common_params) {
   using namespace KokkosSparse;
 
   using mem_space = typename exec_space::memory_space;
@@ -63,26 +60,21 @@ void run_experiment(int argc, char** argv,
   using size_type = default_size_type;
   using lno_t     = default_lno_t;
   using scalar_t  = default_scalar;
-  using crsMat_t =
-      KokkosSparse::CrsMatrix<scalar_t, lno_t, device_t, void, size_type>;
+  using crsMat_t  = KokkosSparse::CrsMatrix<scalar_t, lno_t, device_t, void, size_type>;
 
   using graph_t = typename crsMat_t::StaticCrsGraphType;
 
   LocalParams params;
   if (parse_inputs(params, argc, argv)) return;
 
-  crsMat_t A = KokkosSparse::Impl::read_kokkos_crst_matrix<crsMat_t>(
-      params.mtxFile.c_str());
-  std::cout << "Loaded matrix: " << A.numRows() << "x" << A.numCols()
-            << " with " << A.nnz() << " entries.\n";
+  crsMat_t A = KokkosSparse::Impl::read_kokkos_crst_matrix<crsMat_t>(params.mtxFile.c_str());
+  std::cout << "Loaded matrix: " << A.numRows() << "x" << A.numCols() << " with " << A.nnz() << " entries.\n";
   // This first sort call serves as a warm-up
   KokkosSparse::sort_crs_matrix(A);
-  lno_t m = A.numRows();
-  lno_t n = A.numCols();
-  auto rowmapHost =
-      Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), A.graph.row_map);
-  auto entriesHost =
-      Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), A.graph.entries);
+  lno_t m          = A.numRows();
+  lno_t n          = A.numCols();
+  auto rowmapHost  = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), A.graph.row_map);
+  auto entriesHost = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), A.graph.entries);
   typename crsMat_t::index_type shuffledEntries("shuffled entries", A.nnz());
   // Randomly shuffle the entries within each row, so that the rows aren't
   // already sorted. Leave the values alone; this changes the matrix numerically
@@ -102,13 +94,10 @@ void run_experiment(int argc, char** argv,
     exec.fence();
     totalTime += timer.seconds();
   }
-  std::cout << "Mean sort_crs_matrix time over " << common_params.repeat
-            << " trials: ";
+  std::cout << "Mean sort_crs_matrix time over " << common_params.repeat << " trials: ";
   std::cout << totalTime / common_params.repeat << "\n";
 }
 
 #define KOKKOSKERNELS_PERF_TEST_NAME run_experiment
 #include "KokkosKernels_perf_test_instantiation.hpp"
-int main(int argc, char** argv) {
-  return main_instantiation(argc, argv);
-}  // main
+int main(int argc, char** argv) { return main_instantiation(argc, argv); }  // main