- For sort_crs_matrix/sort_crs_graph, have a heuristic to check if matrix is imbalanced or has long rows (SierraTF use case)
  - For imbalanced/long rows, use a single bulk sort-by-key over all entries. For matrices, this is the same algorithm as sort_bsr_matrix below.
  - Otherwise, use a Kokkos thread-level sort within each row
- For sort_bsr_matrix, use a bulk sort that generates a permutation for all entries and applies it efficiently to the values.
  - avoid having to sort in-place by swapping blocks (expensive)
- Remove sorting functions that have been deprecated for a long time (~2 years!) instead of updating them

Add a simple perf test for sort_crs_matrix too, but as this isn't needed very often, the CMake to enable it is commented out.

| V100 sort_crs_matrix | Before (s)    | After (s)      | Speedup factor |
|----------------------|------------|-------------|----------------|
| Laplace3D_100        | 0.00150869 | 0.000426331 | 3.54           |
| af_shell7            | 0.00153274 | 0.00143674  | 1.07           |
| Hook_1498            | 0.00418005 | 0.00406957  | 1.03           |
| abnormal_energy*     | 0.236816   | 0.0376308   | 6.29           |

*takes the imbalanced/long row path

The new long row implementation could still be improved. If a matrix has a mixture of long and short rows, it would be better
to use the thread-sort for short rows and the bulk (or possibly a team-level) sort for the long rows. The tradeoff is that it's somewhat more complex to implement than what is in this PR.